requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# ML libraries
scikit-learn>=1.3.0
implicit>=0.7.0
scipy>=1.10.0

# MLOps - базовые версии без строгих ограничений
mlflow>=2.8.0
evidently>=0.4.0
great-expectations>=0.18.0

# S3
boto3>=1.34.0
s3fs>=2023.12.0

# Utils
python-dotenv>=1.0.0
joblib>=1.3.0
pyyaml>=6.0.0

# Testing
pytest>=9.0.0

# DVC отдельно - может вызывать конфликты
# dvc[s3]>=3.30.0
